renamed and exported component
