+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './actionItems';
 export * from './alerts';
 export * from './avatar';
 export * from './backdrop';
